requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-qdrant>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
qdrant-client>=1.11.0
